requirements.txt
asgiref==3.8.1
Django==4.2.3
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.7.0
